orders/requirements.txt
Django==3.2.21
djangorestframework==3.14.0
django_rest_passwordreset==1.3.0
drf_spectacular==0.26.5
django-environ==0.11.2
celery==5.3.4